rename and fix ts lint issues
