+++ src/background/initial-assessment-store-data-generator.ts
